+++ templates.go
@@ -32,8 +32,10 @@ func baseFuncMap() template.FuncMap {
 			}
 			return m
 		},
-		"add": func(a, b int) int { return a + b },
-		"sub": func(a, b int) int { return a - b },
+		"add":         func(a, b int) int { return a + b },
+		"sub":         func(a, b int) int { return a - b },
+		"pageNumbers": PageNumbers,
+		"pageURL":     PageURL,
 	}
 }